app/requirements.txt
amqp==2.6.0
aniso8601==8.0.0
attrs==19.3.0
billiard==3.6.3.0
celery==4.4.6
click==7.1.2
Flask==1.1.2
flask-marshmallow==0.13.0
flask-restplus==0.13.0
Flask-SQLAlchemy==2.4.4
future==0.18.2
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
kombu==4.6.11
MarkupSafe==1.1.1
marshmallow==3.7.1
marshmallow-sqlalchemy==0.23.1
psycopg2-binary==2.8.5
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
selenium==3.141.0
six==1.15.0
SQLAlchemy==1.3.18
timeloop==1.0.2
urllib3==1.25.10
uWSGI==2.0.18
vine==1.3.0
Werkzeug==0.16.1
zipp==3.1.0